backend/requirements.txt
fastapi[standard]
psycopg2-binary #postgresql
websockets #for websockets duh
sqlalchemy #for ORM
alembic #for migrations
python-dotenv
